refactor: extract background and witness logic
